Do not force ssh identification to be based on email address
Repo has used the first part of the email address as identification to the
Gerrit server. With this patch it is possible to remove identification all
together or possibly use other values. The value of REPO_SSH_ID will be
inserted into the ssh connection string, if REPO_SSH_ID is empty there
will be no identification at this level, instead ssh should be configured
to identify the user properly.