Update patch set 1
Patch Set 1: No score

That's because it is, honestly. It doesn't address the root problem at all, but I didn't have time to determine what the real cause was and this sounds like it mainly was a traffic growing pains issue for the kernel.org servers with the eclair merge for AOSP, so would solve a current and hopefully temporary pain point.

Patch-set: 1