Fix encoding problem.
I encountered the following error:
│tglase@tglase:~/stuff/android $ repo upload
│Traceback (most recent call last):
│  File "/home/tglase/stuff/android/.repo/repo/main.py", line 235, in <module>
│    _Main(sys.argv[1:])
│  File "/home/tglase/stuff/android/.repo/repo/main.py", line 217, in _Main
│    repo._Run(argv)
│  File "/home/tglase/stuff/android/.repo/repo/main.py", line 123, in _Run
│    cmd.Execute(copts, cargs)
│  File "/home/tglase/stuff/android/.repo/repo/subcmds/upload.py", line 326, in Execute
│    self._MultipleBranches(pending, people)
│  File "/home/tglase/stuff/android/.repo/repo/subcmds/upload.py", line 167, in _MultipleBranches
│    script = Editor.EditString("\n".join(script)).split("\n")
│UnicodeDecodeError: 'ascii' codec can't decode byte 0xc2 in position 32: ordinal not in range(128)

I think this is because I used UTF-8 characters in the commit
messages. This patch fixes at least “repo upload” for me; there
may very well be more places to apply this. I don’t think it
does any harm, seeing as the default was ASCII and UTF-8 is a
superset, but please test this or ask someone who actually
knows Python well.