Modified the submodule implementation
Added a <name> parameter to the manifest --upgrade command. This
will be the projectname on the Gerrit server and will be added to
the .review file.
Add data from the .review file to the config file of the manifest
project during repo init.
Make sure that the manifest project is uploaded last, the manifest
project will create the atomic commit in Gerrit.