Add an option to keep existing obsolete directories on sync
When moving/deleting projects out of the manifest, we may be deleting
projects in which the user has committed, but not uploaded their
changes. This can lead to loss of data.

This change adds an option to repo sync (-k), which will allow users
to sync without deleting the obsolete projects, so they can backup
their commits and port them over to a new project if necessary.