Only sync projects when needed
Use gerrit's ls-projects --show-branch option to sync only projects
which are needed.

The new --show-branch argument of gerrit ls-projects will display the
current HEAD sha-1 for each project in the specified branch. We can
use this command to build a map of the current sha-1 of each project
in the manifest, and compare the revisions we have on disk to the ones
we are working with. This enables us to only sync the projects where
changes were made.