WIP: Only sync projects when needed
This is an attempt to use gerrit's ls-projects --show-branch option to
speed up the syncing of a tree.

I've not yet figured out what I want to do wrt to branches. I think
that the solution I might end up using is to sync the projects in
which the manifest uses a branch different than the "main" one. A
definition of "main" branch might be the branch that the remote uses
by default, since I think that the manifest lets you "override" that
branch, in general.

Interested in feedback on the master ssh stuff.

Change-Id: Ic3242a381c496e080ac827e03ec31c7bb85f28b2