Try fetching a tag manually as a last resort before giving up.
If a tagged commit is not reachable by the fetch refspec configured
for the git (usually refs/heads/*) it will not be downloaded by 'git fetch'.
The tag can however be downloaded with 'git fetch --tags' or
'git fetch tag <tag>'.

This patch fixes the situation when a tag is not found after a
'git fetch'. Repo will issue 'git fetch tag <tag>' before giving
up completely.