init and sync: add --shared=DIR option to share bare repositories
The --shared=DIR option allows multiple android projects to share
the same set of a local repository pool, which contains the super
set of all the individual projects needed by each android project.
It links the .repo/projects to the specified one, instead of
creating its own.

If we have three android projects for different product lines.
And they are composed with different set of projects.

  P1: A B C D
  P2: A B C   E
  P3: A B   D E

The following command sequence will save a lot of traffics:

  $ cd wrk/p1
  $ git init --shared=<somewhere/pool> -u <server> -b <branch_for_p1>
  $ repo sync

  $ cd wrk/p2
  $ git init --shared=<somewhere/pool> -u <server> -b <branch_for_p2>
  $ repo sync

  $ cd wrk/p3
  $ git init --shared=<somewhere/pool> -u <server> -b <branch_for_p3>
  $ repo sync

during developemnt, the 'repo sync' on whichever project will also
benefit to the others.