Update patch set 4
Patch Set 4:

Yes, this functionality is desired, if we can get it to work correctly all of the time.  The idea is that out of 200+ projects that make up the Android platform, only about 5 or so have actually changed since the last time you ran repo sync.  By finding the projects that are different through a single SSH call, we can avoid opening up git fetch for 195 of the projects, saving a lot of round-trip time.

Unfortunately this change is probably stalled.  :-(

Patch-set: 4