Fixed bug identifying 'commit-msg' files.
There was a minor typo that would cause repo to (I believe)
mistakenly identify any file that contained the string 'commit-msg'
as a commit message hook.  I believe that it was intended that repo
only recognize files named exactly 'commit-msg'.

If the substring behavior was desired, let me know and I'll
resubmit a patch that removes the parenthesis around ('commit-msg').