Add support for repo-level pre-upload hook and prepare for future hooks.
All repo-level hooks are expected to live in a single project at the top level of that project.  The name of the hooks project is provided in the manifest.xml.  The manifest also lists which hooks are enabled to make it obvious if a file somehow failed to sync down (or got deleted).

Before running any hook, we will prompt the user to make sure that it is OK.  A user can deny running the hook, allow once, or allow "forever" (until hooks change).  This tries to keep with the git spirit of not automatically running anything on the user's computer that got synced down.  Note that individual repo commands can add always options to avoid these prompts as they see fit (see below for the 'upload' options).

When hooks are run, they are loaded into the current interpreter (the one running repo) and their main() function is run.  This mechanism is used (instead of using subprocess) to make it easier to expand to a richer hook interface in the future.  During loading, the interpreter's sys.path is updated to contain the directory containing the hooks so that hooks can be split into multiple files.

The upload command has two options that control hook behavior: '--no-verify' and '--verify'.
  - no-verify=False, verify=False (DEFAULT):
    If stdout is a tty, can prompt about running upload hooks if needed.
    If user denies running hooks, the upload is cancelled.  If stdout is
    not a tty and we would need to prompt about upload hooks, upload is
    cancelled.
  - no-verify=False, verify=True:
    Always run upload hooks with no prompt.
  - no-verify=True, verify=False:
    Never run upload hooks, but upload anyway (AKA bypass hooks).
  - no-verify=True, verify=True:
    Invalid

Change-Id: I5cefa2cd5865c72589263cf8e2f152a43c122f70