Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Anush:

Here's a patch you can play with for checkout out the '--depth' stuff.  I will admit that I don't have lots of confidence in this patch / background knowledge of it, so I've marked it -1.  It's just hacking the --depth into the fetch command, as you requested.

To test it, I did a 'repo init' with a bogus URL (to get repo down).  Then, I applied this patch to repo.  Then, I did a repo init with the proper URL and a --depth=1.

On a minilayout, I noticed a 230MB savings.

A few specific weirdnesses:

* The depth setting is stored in the git settings of the manifest, for lack of a better place.

* If you end up changing your depth through a later 'repo init', it will only take effect for _new_ things synced down.  I don't think we end up re-running the 'fetch' command for old projects that were synced.  I didn't look into this.

* If you do a repo init once with a --depth, then later do a repo init without a --depth, that will be interpreted as meaning that any future projects should be synced with unlimited depth (rather than being interpreted as leaving the depth alone).

Patch-set: 1