Made main repo script python3 compatible.
This is the first in a series of commits to make repo python3
compatibile.  Through the use of __future__ and some clever duck-punching,
we can make valid python3 code that will run in python2.  When python2
is no longer supported we will be able to just remove the block of code
that imports from __future__ and does the duck-punching.

Change-Id: I94f1176f2e9b5573b3b6e9ef0be086f6a1ff876b