repo forall will warn (and can exit) on command errors
The 'forall' command has previously continued without any warning
if a command (as specified with -c) exited with non-zero status
although the 'forall' command itself would exit with non-zero
status.

The default behavior is now changed so that it will print out
warnings on stderr after every command that exited unsuccessfully
and it will also print a warning at the end of the 'forall' command.

This legacy behavior can still be used by specifying the
--discard-warnings option on the command line.

If -e is specified, the 'forall' command will exit immediately
after a command has exited unsuccessfully and will not iterate
through the remaining projects.

Bug: Issue 17
Change-Id: Ibea405e0d98b575ad3bda719d511f6982511c19c
Signed-off-by: Victor Boivie <victor.boivie@sonyericsson.com>